split to selectors.ts and utils.ts
