+++ ui/components/app/assets/nfts/nfts-detection-notice-nfts-tab/nfts-detection-notice-nfts-tab.js
@@ -8,7 +8,7 @@ import {
   setUseNftDetection,
 } from '../../../../../store/actions';
 import { getOpenSeaEnabled } from '../../../../../selectors';
-import { setShowNftDetectionEnablementToast } from '../../../toast-master/toast-master-selectors';
+import { setShowNftDetectionEnablementToast } from '../../../toast-master/selectors';
 
 export default function NFTsDetectionNoticeNFTsTab() {
   const t = useI18nContext();
